Move behavior and enqueuedby to isolated files
