+++ owncloudApp/src/main/java/com/owncloud/android/usecases/UploadFileUseCase.kt
@@ -148,54 +148,3 @@ class UploadFileUseCase(
         val wifiOnly: Boolean,
     )
 }
-
-/**
- * Analog to the old LOCAL_BEHAVIOUR but with fixed options.
- *
- * By default, the file will be copied. We do not want to remove the file without user approval.
- *
- * Warning -> Order of elements is really important. The ordinal is used to store the value in the database.
- */
-enum class UploadBehavior {
-    COPY, MOVE;
-
-    companion object {
-        fun fromLegacyLocalBehavior(oldLocalBehavior: Int): UploadBehavior {
-            return when (oldLocalBehavior) {
-                FileUploader.LOCAL_BEHAVIOUR_MOVE -> MOVE
-                FileUploader.LOCAL_BEHAVIOUR_COPY -> COPY
-                else -> COPY
-            }
-        }
-    }
-}
-
-/**
- * Analog to the old CREATED_BY but with fixed options.
- *
- * By default, we consider that it is enqueued manually.
- *
- * Warning -> Order of elements is really important. The ordinal is used to store the value in the database.
- */
-enum class UploadEnqueuedBy {
-    ENQUEUED_BY_USER,
-    ENQUEUED_AS_CAMERA_UPLOAD_PICTURE,
-    ENQUEUED_AS_CAMERA_UPLOAD_VIDEO;
-
-    fun fromLegacyCreatedBy(oldCreatedBy: Int): UploadEnqueuedBy {
-        return when (oldCreatedBy) {
-            CREATED_BY_USER -> ENQUEUED_BY_USER
-            CREATED_AS_CAMERA_UPLOAD_PICTURE -> ENQUEUED_AS_CAMERA_UPLOAD_PICTURE
-            CREATED_AS_CAMERA_UPLOAD_VIDEO -> ENQUEUED_AS_CAMERA_UPLOAD_VIDEO
-            else -> ENQUEUED_BY_USER
-        }
-    }
-
-    fun toTransferTag(): String {
-        return when (this) {
-            ENQUEUED_BY_USER -> TRANSFER_TAG_MANUAL_UPLOAD
-            ENQUEUED_AS_CAMERA_UPLOAD_PICTURE -> TRANSFER_TAG_CAMERA_UPLOAD
-            ENQUEUED_AS_CAMERA_UPLOAD_VIDEO -> TRANSFER_TAG_CAMERA_UPLOAD
-        }
-    }
-}
